resources are now their respective own project
AboutBox moved to LogExpert.UI
