+++ src/Directory.Packages.props
@@ -28,6 +28,8 @@
     <PackageVersion Include="NUnit3TestAdapter" Version="5.0.0" />
     <PackageVersion Include="SSH.NET" Version="2024.2.0" />
     <PackageVersion Include="System.Data.DataSetExtensions" Version="4.5.0" />
+    <PackageVersion Include="System.Drawing.Common" Version="9.0.3" />
+    <PackageVersion Include="System.Resources.Extensions" Version="9.0.3" />
     <PackageVersion Include="System.Text.Encoding" Version="4.3.0" />
   </ItemGroup>
 </Project>
