+++ src/LogExpert.sln
@@ -57,9 +57,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RegexColumnizer.UnitTests",
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LogexpertgRPCService", "LogexpertgRPCService\LogexpertgRPCService.csproj", "{D31D1721-9DEA-45A4-B813-D5023B2EC2CC}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Logexpert.Core", "Logexpert.Core\Logexpert.Core.csproj", "{F49C6738-3F62-4890-8FF2-1F53A0F0A2CD}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LogExpert.Core", "Logexpert.Core\LogExpert.Core.csproj", "{F49C6738-3F62-4890-8FF2-1F53A0F0A2CD}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Logexpert.Resources", "Logexpert.Resources\Logexpert.Resources.csproj", "{578CC5D5-1DCD-47C2-8BD3-B32C14635BEC}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LogExpert.Resources", "Logexpert.Resources\LogExpert.Resources.csproj", "{578CC5D5-1DCD-47C2-8BD3-B32C14635BEC}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "LogExpert.UI", "Logexpert.UI\LogExpert.UI.csproj", "{5DB0B0F8-3C3E-4805-93A3-7E69AC45CD40}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -145,6 +147,10 @@ Global
 		{578CC5D5-1DCD-47C2-8BD3-B32C14635BEC}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{578CC5D5-1DCD-47C2-8BD3-B32C14635BEC}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{578CC5D5-1DCD-47C2-8BD3-B32C14635BEC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5DB0B0F8-3C3E-4805-93A3-7E69AC45CD40}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5DB0B0F8-3C3E-4805-93A3-7E69AC45CD40}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5DB0B0F8-3C3E-4805-93A3-7E69AC45CD40}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5DB0B0F8-3C3E-4805-93A3-7E69AC45CD40}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
