+++ src/LogExpert/Controls/LogTabWindow/LogTabWindow.cs
@@ -1,7 +1,6 @@
 ﻿using LogExpert.Config;
 using LogExpert.Dialogs;
 using LogExpert.Entities;
-using LogExpert.Entities.EventArgs;
 using LogExpert.Extensions.Forms;
 using LogExpert.Interface;
 
@@ -150,7 +149,7 @@ namespace LogExpert.Controls.LogTabWindow
             // get a list of resource names from the manifest
             string[] resNames = a.GetManifestResourceNames();
 
-            Bitmap bmp = Properties.Resources.Deceased;
+            Bitmap bmp = Resources.Resources.Deceased;
             _deadIcon = Icon.FromHandle(bmp.GetHicon());
             bmp.Dispose();
             Closing += OnLogTabWindowClosing;
@@ -239,7 +238,7 @@ namespace LogExpert.Controls.LogTabWindow
             // Tabs line
             dockPanel.Theme.Skin.DockPaneStripSkin.ToolWindowGradient.DockStripGradient.StartColor = ColorMode.TabsBackgroundStripColor;
             dockPanel.Theme.Skin.DockPaneStripSkin.ToolWindowGradient.DockStripGradient.EndColor = ColorMode.TabsBackgroundStripColor;
-            
+
             dockPanel.Theme.Skin.DockPaneStripSkin.DocumentGradient.DockStripGradient.StartColor = ColorMode.TabsBackgroundStripColor;
             dockPanel.Theme.Skin.DockPaneStripSkin.DocumentGradient.DockStripGradient.EndColor = ColorMode.TabsBackgroundStripColor;
 
@@ -247,11 +246,11 @@ namespace LogExpert.Controls.LogTabWindow
             dockPanel.Theme.Skin.DockPaneStripSkin.ToolWindowGradient.ActiveTabGradient.StartColor = ColorMode.ActiveTabColor;
             dockPanel.Theme.Skin.DockPaneStripSkin.ToolWindowGradient.ActiveTabGradient.EndColor = ColorMode.ActiveTabColor;
             dockPanel.Theme.Skin.DockPaneStripSkin.ToolWindowGradient.ActiveTabGradient.TextColor = ColorMode.ForeColor;
-            
+
             dockPanel.Theme.Skin.DockPaneStripSkin.DocumentGradient.ActiveTabGradient.StartColor = ColorMode.ActiveTabColor;
             dockPanel.Theme.Skin.DockPaneStripSkin.DocumentGradient.ActiveTabGradient.EndColor = ColorMode.ActiveTabColor;
             dockPanel.Theme.Skin.DockPaneStripSkin.DocumentGradient.ActiveTabGradient.TextColor = ColorMode.ForeColor;
-            
+
             dockPanel.Theme.Skin.DockPaneStripSkin.DocumentGradient.InactiveTabGradient.StartColor = ColorMode.InactiveTabColor;
             dockPanel.Theme.Skin.DockPaneStripSkin.DocumentGradient.InactiveTabGradient.EndColor = ColorMode.InactiveTabColor;
             dockPanel.Theme.Skin.DockPaneStripSkin.DocumentGradient.InactiveTabGradient.TextColor = ColorMode.ForeColor;
