+++ src/LogExpert/Controls/LogWindow/LogWindow.cs
@@ -228,8 +228,8 @@ namespace LogExpert.Controls.LogWindow
             dataGridView.EditModeMenuStrip = editModeContextMenuStrip;
             markEditModeToolStripMenuItem.Enabled = true;
 
-            _panelOpenButtonImage = Properties.Resources.Arrow_menu_open;
-            _panelCloseButtonImage = Properties.Resources.Arrow_menu_close;
+            _panelOpenButtonImage = Resources.Resources.Arrow_menu_open;
+            _panelCloseButtonImage = Resources.Resources.Arrow_menu_close;
 
             Settings settings = ConfigManager.Settings;
