+++ src/LogExpert/Dialogs/AboutBox.cs
@@ -1,132 +1,128 @@
-﻿using System.Diagnostics;
-using System.IO;
-using System.Reflection;
-using System.Windows.Forms;
+﻿//using System.Diagnostics;
+//using System.IO;
+//using System.Reflection;
+//using System.Windows.Forms;
 
-namespace LogExpert.Dialogs
-{
-    internal partial class AboutBox : Form
-    {
-        #region Fields
+//namespace LogExpert.Dialogs
+//{
+//    internal partial class AboutBox : Form
+//    {
+//        #region Fields
 
-        private readonly Assembly _assembly;
+//        private readonly Assembly _assembly;
 
-        #endregion
+//        #endregion
 
-        #region cTor
+//        #region cTor
 
-        public AboutBox()
-        {
-            InitializeComponent();
+//        public AboutBox()
+//        {
+//            InitializeComponent();
 
-            _assembly = Assembly.GetExecutingAssembly();
+//            _assembly = Assembly.GetExecutingAssembly();
 
-            Text = $@"About {AssemblyTitle}";
-            labelProductName.Text = AssemblyProduct;
-            labelVersion.Text = AssemblyVersion;
-            labelCopyright.Text = AssemblyCopyright;
-            //this.labelCompanyName.Text = AssemblyCompany;
-            textBoxDescription.Text = AssemblyDescription +
-                                           "\r\n\r\nCredits:\r\n\r\n" +
-                                           "DockPanel control (c) 2007 Weifen Luo \r\n" +
-                                           "Early bird test: Mathias Dräger\r\n" +
-                                           "\r\n" +
-                                           "LogExpert uses modules from:\r\n" +
-                                           "http://sourceforge.net/projects/dockpanelsuite/\r\n" +
-                                           "http://www.xml-rpc.net/";
-            string link = "https://github.com/LogExperts/LogExpert";
-            linkLabelURL.Links.Add(new LinkLabel.Link(0, link.Length, link));
-        }
+//            Text = $@"About {AssemblyTitle}";
+//            labelProductName.Text = AssemblyProduct;
+//            labelVersion.Text = AssemblyVersion;
+//            labelCopyright.Text = AssemblyCopyright;
+//            //this.labelCompanyName.Text = AssemblyCompany;
+//            textBoxDescription.Text = AssemblyDescription +
+//                                           "\r\n\r\nCredits:\r\n\r\n" +
+//                                           "LogExpert uses modules from:\r\n" +
+//                                           "https://github.com/dockpanelsuite/dockpanelsuite\r\n";
+//            string link = "https://github.com/LogExperts/LogExpert";
+//            linkLabelURL.Links.Add(new LinkLabel.Link(0, link.Length, link));
+//        }
 
-        #endregion
+//        #endregion
 
-        #region Properties
+//        #region Properties
 
-        public string AssemblyTitle
-        {
-            get
-            {
-                object[] attributes = _assembly.GetCustomAttributes(typeof(AssemblyTitleAttribute), false);
-                if (attributes.Length > 0)
-                {
-                    AssemblyTitleAttribute titleAttribute = (AssemblyTitleAttribute)attributes[0];
-                    if (titleAttribute.Title != "")
-                    {
-                        return titleAttribute.Title;
-                    }
-                }
-                return Path.GetFileNameWithoutExtension(_assembly.Location);
-            }
-        }
+//        public string AssemblyTitle
+//        {
+//            get
+//            {
+//                object[] attributes = _assembly.GetCustomAttributes(typeof(AssemblyTitleAttribute), false);
+//                if (attributes.Length > 0)
+//                {
+//                    AssemblyTitleAttribute titleAttribute = (AssemblyTitleAttribute)attributes[0];
+//                    if (titleAttribute.Title != string.Empty)
+//                    {
+//                        return titleAttribute.Title;
+//                    }
+//                }
+//                return Path.GetFileNameWithoutExtension(_assembly.Location);
+//            }
+//        }
 
-        public string AssemblyVersion
-        {
-            get
-            {
-                AssemblyName assembly = _assembly.GetName();
-                string version = $"{assembly.Version.Major}.{assembly.Version.Minor}.{assembly.Version.Build}.{assembly.Version.Revision}";
-                if (assembly.Version.Revision >= 9000)
-                {
-                    version += " Testrelease";
-                }
+//        public string AssemblyVersion
+//        {
+//            get
+//            {
+//                AssemblyName assembly = _assembly.GetName();
+//                string version = $"{assembly.Version.Major}.{assembly.Version.Minor}.{assembly.Version.Build}.{assembly.Version.Revision}";
+//                if (assembly.Version.Revision >= 9000)
+//                {
+//                    version += " Testrelease";
+//                }
 
-                return version;
-            }
-        }
+//                return version;
+//            }
+//        }
 
-        public string AssemblyDescription
-        {
-            get
-            {
-                object[] attributes = _assembly.GetCustomAttributes(typeof(AssemblyDescriptionAttribute), false);
-                if (attributes.Length == 0)
-                {
-                    return string.Empty;
-                }
-                return ((AssemblyDescriptionAttribute)attributes[0]).Description;
-            }
-        }
+//        public string AssemblyDescription
+//        {
+//            get
+//            {
+//                object[] attributes = _assembly.GetCustomAttributes(typeof(AssemblyDescriptionAttribute), false);
+//                if (attributes.Length == 0)
+//                {
+//                    return string.Empty;
+//                }
+//                return ((AssemblyDescriptionAttribute)attributes[0]).Description;
+//            }
+//        }
 
-        public string AssemblyProduct
-        {
-            get
-            {
-                object[] attributes = _assembly.GetCustomAttributes(typeof(AssemblyProductAttribute), false);
-                if (attributes.Length == 0)
-                {
-                    return string.Empty;
-                }
-                return ((AssemblyProductAttribute)attributes[0]).Product;
-            }
-        }
+//        public string AssemblyProduct
+//        {
+//            get
+//            {
+//                object[] attributes = _assembly.GetCustomAttributes(typeof(AssemblyProductAttribute), false);
+//                if (attributes.Length == 0)
+//                {
+//                    return string.Empty;
+//                }
+//                return ((AssemblyProductAttribute)attributes[0]).Product;
+//            }
+//        }
 
-        public string AssemblyCopyright
-        {
-            get
-            {
-                object[] attributes = _assembly.GetCustomAttributes(typeof(AssemblyCopyrightAttribute), false);
-                if (attributes.Length == 0)
-                {
-                    return string.Empty;
-                }
-                return ((AssemblyCopyrightAttribute)attributes[0]).Copyright;
-            }
-        }
+//        public string AssemblyCopyright
+//        {
+//            get
+//            {
+//                object[] attributes = _assembly.GetCustomAttributes(typeof(AssemblyCopyrightAttribute), false);
+//                if (attributes.Length == 0)
+//                {
+//                    return string.Empty;
+//                }
+//                return ((AssemblyCopyrightAttribute)attributes[0]).Copyright;
+//            }
+//        }
 
-        #endregion
+//        #endregion
 
-        #region Events handler
+//        #region Events handler
 
-        private void OnLinkLabelURLClicked(object sender, LinkLabelLinkClickedEventArgs e)
-        {
-            string target = e.Link.LinkData as string;
-            Process.Start(new ProcessStartInfo
-            {
-                UseShellExecute = true,
-                FileName = target,
-            });
-        }
+//        private void OnLinkLabelURLClicked(object sender, LinkLabelLinkClickedEventArgs e)
+//        {
+//            string target = e.Link.LinkData as string;
+//            Process.Start(new ProcessStartInfo
+//            {
+//                UseShellExecute = true,
+//                FileName = target,
+//            });
+//        }
 
-        #endregion
-    }
-}
+//        #endregion
+//    }
+//}
